+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_border_all/avoid_border_all_rule.dart
@@ -1,0 +1,68 @@
+// ignore_for_file: public_member_api_docs
+
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
+
+import '../../../../../utils/node_utils.dart';
+import '../../../lint_utils.dart';
+import '../../../models/internal_resolved_unit_result.dart';
+import '../../../models/issue.dart';
+import '../../../models/replacement.dart';
+import '../../../models/severity.dart';
+import '../../models/flutter_rule.dart';
+import '../../rule_utils.dart';
+
+part 'visitor.dart';
+
+class AvoidBorderAllRule extends FlutterRule {
+  static const ruleId = 'avoid-border-all';
+  static const _issueMessage =
+      'Prefer using const constructor Border.fromBorderSide.';
+  static const _replaceComment = 'Replace with const Border.fromBorderSide.';
+
+  AvoidBorderAllRule([Map<String, Object> config = const {}])
+      : super(
+          id: ruleId,
+          severity: readSeverity(config, Severity.performance),
+          excludes: readExcludes(config),
+        );
+
+  @override
+  Iterable<Issue> check(InternalResolvedUnitResult source) {
+    final visitor = _Visitor();
+    source.unit.visitChildren(visitor);
+
+    return visitor.expressions
+        .map((expression) => createIssue(
+              rule: this,
+              location: nodeLocation(
+                node: expression,
+                source: source,
+                withCommentOrMetadata: true,
+              ),
+              message: _issueMessage,
+              replacement: _createReplacement(expression),
+            ))
+        .toList(growable: false);
+  }
+
+  Replacement? _createReplacement(InstanceCreationExpression expression) {
+    final value = _getConstructorArgumentValue(expression);
+
+    return Replacement(
+      comment: _replaceComment,
+      replacement: 'const Border.fromBorderSide(BorderSide($value))',
+    );
+  }
+
+  String? _getConstructorArgumentValue(InstanceCreationExpression expression) {
+    final arguments = expression.argumentList.arguments;
+    final buffer = StringBuffer();
+
+    for (final arg in arguments) {
+      buffer.write('${arg.toString()},');
+    }
+
+    return buffer.toString();
+  }
+}